Abzu-180207-Update NVDSuppressions.xml notes for JScript CVE
Enhanced the notes section in NVDSuppressions.xml to provide a detailed explanation of the CVE references. The updated notes clarify that the CVE pertains to the Microsoft Chakra JavaScript, JScript, and VBScript engines used in Internet Explorer 10 and 11 and Microsoft Edge. The suppression is due to a false positive for this CVE in JScript.